requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
